+++ react/PickupPackages.tsx
@@ -7,15 +7,7 @@ import StorePickUpHeader from './components/PickUpHeader'
 
 const CSS_HANDLES = ['package', 'packageHeaderColumn']
 
-interface Props {
-  displayTitle?: boolean
-  displayProducts?: boolean
-}
-
-const PickupPackages: FC<Props> = ({
-  displayTitle = false,
-  displayProducts = true,
-}) => {
+const PickupPackages: FC = () => {
   const handles = useCssHandles(CSS_HANDLES)
   const { pickUpParcels } = useOrder()
 
@@ -38,10 +30,9 @@ const PickupPackages: FC<Props> = ({
               shippingData={pickup}
               index={index}
               numPackages={pickUpParcels.length}
-              displayTitle={displayTitle}
             />
           </div>
-          {displayProducts ? <ProductList products={pickup.items} /> : ''}
+          <ProductList products={pickup.items} />
         </div>
       ))}
     </Fragment>