+++ react/components/DeliveryHeader.tsx
@@ -9,7 +9,6 @@ interface Props {
   index: number
   numPackages: number
   giftRegistry?: GiftRegistry | null
-  displayTitle?: boolean
 }
 
 const CSS_HANDLES = [
@@ -27,7 +26,6 @@ const DeliveryHeader: FC<Props> = ({
   index,
   numPackages,
   giftRegistry,
-  displayTitle = false,
 }) => {
   const handles = useCssHandles(CSS_HANDLES)
   const multipleDeliveries = numPackages > 1
@@ -68,13 +66,9 @@ const DeliveryHeader: FC<Props> = ({
       {giftRegistry &&
       giftRegistry.addressId === shippingData.address.addressId ? (
         <div className={`${handles.packageGiftDescription} c-muted-1`}>
-          {displayTitle ? (
-            <span className={`${handles.packageAddressTitle}`}>
-              <FormattedMessage id="store/shipping.header.address" />
-            </span>
-          ) : (
-            ''
-          )}
+          <span className={`${handles.packageAddressTitle} dn`}>
+            <FormattedMessage id="store/shipping.header.address" />
+          </span>
           <FormattedMessage
             id="store/shipping.header.wishlist.address"
             values={{ giftRegistryName: giftRegistry.description }}
@@ -82,13 +76,9 @@ const DeliveryHeader: FC<Props> = ({
         </div>
       ) : (
         <div className={`${handles.packageAddressWrapper} mb5 mr10-m`}>
-          {displayTitle ? (
-            <span className={`${handles.packageAddressTitle}`}>
-              <FormattedMessage id="store/shipping.header.address" />
-            </span>
-          ) : (
-            ''
-          )}
+          <span className={`${handles.packageAddressTitle} dn`}>
+            <FormattedMessage id="store/shipping.header.address" />
+          </span>
           <Address address={shippingData.address} />
         </div>
       )}