+++ react/components/PickUpHeader.tsx
@@ -8,7 +8,6 @@ interface Props {
   shippingData: Parcel
   index: number
   numPackages: number
-  displayTitle?: boolean
 }
 
 const CSS_HANDLES = [
@@ -18,14 +17,10 @@ const CSS_HANDLES = [
   'packageAddressWrapper',
   'packageReceiver',
   'packageAdditionalInfo',
-]
+  'packageAddressTitle',
+] as const
 
-const StorePickUpHeader: FC<Props> = ({
-  shippingData,
-  index,
-  numPackages,
-  displayTitle = false,
-}) => {
+const StorePickUpHeader: FC<Props> = ({ shippingData, index, numPackages }) => {
   const handles = useCssHandles(CSS_HANDLES)
   const multiplePickups = numPackages > 1
   const { receiverName } = shippingData.address
@@ -61,13 +56,9 @@ const StorePickUpHeader: FC<Props> = ({
       <div
         className={`${handles.packageInfoWrapper} flex justify-left-m flex-column-s flex-column-l flex-row-m justify-between-l justify-start-l`}
       >
-        {displayTitle ? (
-          <span className={`${handles.packageAddressTitle}`}>
-            <FormattedMessage id="store/shipping.header.address" />
-          </span>
-        ) : (
-          ''
-        )}
+        <span className={`${handles.packageAddressTitle} dn`}>
+          <FormattedMessage id="store/shipping.header.address" />
+        </span>
         <div className={`${handles.packageAddressWrapper} mb5 mr10-m`}>
           <Address address={shippingData.address} pickup={shippingData} />
         </div>